refactor(dubbo-samples-configcenter-api): upgrade samples
